app/src/main/kotlin/com/dg/eqs/core/definition/term/operation/alteration/condensingstep/extractor/DashOperationAndAnythingExtractor.kt
package com.dg.eqs.core.definition.term.operation.alteration.condensingstep.extractor

import com.dg.eqs.base.abbreviation.termsOf
import com.dg.eqs.core.definition.term.Terms
import com.dg.eqs.core.definition.term.operation.alteration.condensingstep.CondensingExtractor
import com.dg.eqs.core.definition.term.operation.alteration.condensingstep.and
import com.dg.eqs.core.definition.term.operation.dashoperation.DashOperation


object DashOperationAndAnythingExtractor : CondensingExtractor {
    override fun extract(operands: Terms) = with(operands) {
        val indexOfFirstDashOperation = indexOf<DashOperation>()

        val left = termsOf(get(indexOfFirstDashOperation))
        val right = termsOf(get(indexOfFirstDashOperation + 1))

        left and right
    }
}